+++ apstools/devices/tests/test_area_detector_support.py
@@ -14,78 +14,188 @@ import pytest
 from bluesky import plan_stubs as bps
 from bluesky import plans as bp
 
+from ...tests import IOC_AD
 from ...tests import READ_PATH_TEMPLATE
+from ...tests import WRITE_PATH_TEMPLATE
 from ...tests import MonitorCache
+from ...tests import in_gha_workflow
 from ...tests import timed_pause
 from .. import AD_EpicsFileNameHDF5Plugin
 from .. import AD_EpicsFileNameJPEGPlugin
 from .. import AD_EpicsFileNameMixin
 from .. import AD_EpicsFileNameTIFFPlugin
 from .. import AD_full_file_name_local
+from .. import AD_plugin_primed
+from .. import AD_prime_plugin2
 from .. import BadPixelPlugin
+from .. import ad_creator
+
+
+def caput_AD(signal, value):
+    """Low-level caput to AD Component (avoids failing Status objects)."""
+    import epics
+
+    epics.caput(signal._write_pv.pvname, value)
+
+
+def custom_ad_creator(plugin_name):
+    """Create area detector with named file writer plugin."""
+    plugin_class = dict(
+        hdf1=AD_EpicsFileNameHDF5Plugin,
+        jpeg1=AD_EpicsFileNameJPEGPlugin,
+        tiff1=AD_EpicsFileNameTIFFPlugin,
+    )
+    writer_plugin = {
+        plugin_name: {
+            "class": plugin_class[plugin_name],
+            "write_path_template": WRITE_PATH_TEMPLATE,
+            "read_path_template": READ_PATH_TEMPLATE,
+        }
+    }
+    if plugin_name in "jpeg1 tiff1".split():
+        writer_plugin[plugin_name]["suffix"] = f"{plugin_name.upper()}:"
+    adsimdet = ad_creator(
+        IOC_AD,
+        name="adsimdet",
+        class_name="MySimDet",
+        plugins=[
+            "cam",
+            "image",
+            "pva",
+            writer_plugin,
+        ],
+    )
+    return adsimdet
+
+
+def unprime_AD_plugin(plugin):
+    """
+    Make area detector plugin require priming again.
+
+    One way is to prime the plugin with a different data type.
+    """
+    cam = plugin.parent.cam
+    original_type = cam.data_type.get(use_monitor=False, as_string=True)
+    assert original_type != "Int8"
+
+    cam.data_type.put("Int8")  # change it
+    AD_prime_plugin2(plugin)
+    timed_pause(0.1)
+    assert AD_plugin_primed(plugin)
+
+    cam.data_type.put(original_type)  # restore it
+    timed_pause(0.1)
+    assert not AD_plugin_primed(plugin)
 
 
 @pytest.mark.parametrize(
-    # fmt: off
     "plugin_name, spec",
     [
         ["hdf1", "AD_HDF5"],
         ["jpeg1", "AD_JPEG"],
         ["tiff1", "AD_TIFF"],
     ],
-    # fmt: on
 )
-def test_AD_EpicsFileNameMixin(plugin_name, spec, adsimdet):
-    assert adsimdet is not None
-    assert plugin_name in dir(adsimdet)
+@pytest.mark.parametrize(
+    "presets, prime, context, expected",
+    [
+        [
+            dict(
+                auto_save=0,  # "No"
+                lazy_open="No",
+                create_directory=0,
+                file_write_mode=2,  # "Single"
+                file_template="%s%s_%d.trouble",
+                file_path="/tmp/",
+                file_name="test",
+                file_number=1 + int(9 * random.random()),
+            ),
+            False,
+            does_not_raise(),
+            None,
+        ],
+        [
+            dict(
+                auto_save=0,  # "No"
+                lazy_open="No",
+                create_directory=0,
+                file_template="",
+                file_path="",
+                file_name="",
+                file_number=1 + int(9 * random.random()),
+            ),
+            True,
+            pytest.raises(ValueError),
+            ":FileName' is an empty string.",
+        ],
+        [
+            dict(
+                auto_save=0,  # "No"
+                lazy_open="No",
+                create_directory=-5,
+                file_template="%s%s_%2.2d.ad_image",
+                file_path="/tmp/",
+                file_name="test",
+                file_number=1 + int(9 * random.random()),
+            ),
+            True,
+            does_not_raise(),
+            None,
+        ],
+    ],
+)
+def test_AD_EpicsFileNameMixin(plugin_name, spec, presets, prime, context, expected):
+    """Test the mixin with stage/unstage."""
+    with context as exinfo:
+        det = custom_ad_creator(plugin_name)
 
-    # basic plugin tests
-    mixin = getattr(adsimdet, plugin_name)
-    assert AD_EpicsFileNameMixin is not None
-    assert isinstance(mixin, AD_EpicsFileNameMixin)
-    assert "filestore_spec" in dir(mixin)
-    assert mixin.filestore_spec == spec
-    assert isinstance(mixin.stage_sigs, dict)
-    assert len(mixin.stage_sigs) >= 4
-    assert "capture" in mixin.stage_sigs
+        plugin = getattr(det, plugin_name)
+        assert AD_EpicsFileNameMixin is not None
+        assert isinstance(plugin, AD_EpicsFileNameMixin)
+        assert "filestore_spec" in dir(plugin)
+        assert plugin.filestore_spec == spec
+        assert isinstance(plugin.stage_sigs, dict)
+        assert len(plugin.stage_sigs) >= 4
+        assert plugin.stage_sigs.get("file_write_mode") == "Stream"
+        assert plugin.stage_sigs.get("capture") in (1, "Capture")
 
-    # configuration prescribed for the user
-    user_settings = dict(
-        array_counter=0,
-        auto_increment=1,  # Yes
-        auto_save=1,  # Yes
-        create_directory=-5,
-        file_name="flotsam",
-        file_number=1 + int(10 * random.random()),
-        file_path=f"{pathlib.Path(tempfile.mkdtemp())}/",  # ALWAYS ends with "/"
-        file_template=f"%s%s_%2.2d.{plugin_name}",
-        num_capture=1 + int(10 * random.random()),
-    )
-    if plugin_name == "hdf1":
-        user_settings["compression"] = "zlib"
+        plugin.stage_sigs.move_to_end("capture", last=True)
+        assert list(plugin.stage_sigs)[-1] == "capture"
 
-    # add the settings
-    for k, v in user_settings.items():
-        getattr(mixin, k).put(v)
-    timed_pause()
+        unprime_AD_plugin(plugin)
 
-    adsimdet.stage()
-    if plugin_name == "hdf1":  # Why special case?
-        user_settings["file_number"] += 1  # the IOC will do the same
-    assert list(mixin.stage_sigs.keys())[-1] == "capture"
-    for k, v in user_settings.items():
-        assert getattr(mixin, k).get() == v, f"{plugin_name=} {k=}  {v=}"
+        if prime and not AD_plugin_primed(plugin):
+            AD_prime_plugin2(plugin)
+            assert AD_plugin_primed(plugin)
 
-    assert mixin.get_frames_per_point() == user_settings["num_capture"]
+        settings = presets.copy()  # Do NOT change the 'presets' parameter.
+        if plugin_name == "hdf1":
+            settings["compression"] = "zlib"
 
-    filename, read_path, write_path = mixin.make_filename()
-    assert isinstance(filename, str)
-    assert isinstance(read_path, str)
-    assert isinstance(write_path, str)
-    assert filename == user_settings["file_name"]
-    assert read_path.startswith(datetime.datetime.now().strftime(READ_PATH_TEMPLATE))
-    assert write_path == user_settings["file_path"]
-    adsimdet.unstage()
+        for attr, value in settings.items():
+            if attr in dir(plugin):
+                caput_AD(getattr(plugin, attr), value)
+
+        try:
+            det.stage()
+
+            assert list(plugin.stage_sigs.keys())[-1] == "capture"
+            for k, v in settings.items():
+                assert getattr(plugin, k).get(use_monitor=False) == v, f"{plugin_name=} {k=}  {v=}"
+
+            filename, read_path, write_path = plugin.make_filename()
+            assert isinstance(filename, str)
+            assert isinstance(read_path, str)
+            assert isinstance(write_path, str)
+            assert filename == settings["file_name"]
+            assert read_path.startswith(datetime.datetime.now().strftime(READ_PATH_TEMPLATE))
+            assert write_path == settings["file_path"]
+
+        finally:
+            det.unstage()
+
+    if expected is not None:
+        assert expected in str(exinfo)
 
 
 @pytest.mark.parametrize(
@@ -121,8 +231,10 @@ def test_stage(plugin_name, plugin_class, adsimdet, fname):
         ["tiff1", AD_EpicsFileNameTIFFPlugin],
     ],
 )
-def test_bp_count_custom_name(plugin_name, plugin_class, adsimdet, fname):
+def test_bp_count_custom_name(plugin_name, plugin_class, fname):
     assert isinstance(plugin_name, str)
+
+    adsimdet = custom_ad_creator(plugin_name)
 
     plugin = getattr(adsimdet, plugin_name)
     assert isinstance(plugin, plugin_class)
@@ -142,7 +254,7 @@ def test_bp_count_custom_name(plugin_name, plugin_class, adsimdet, fname):
             adsimdet.cam.acquire_time, 0.001,
             adsimdet.cam.acquire_period, 0.002,
             adsimdet.cam.image_mode, "Single",
-            # plugin.file_template  # pre-configured
+            plugin.file_template, "%s%s_%3.3d.h5"
         )
         # fmt: on
 
@@ -189,7 +301,8 @@ def test_bp_count_custom_name(plugin_name, plugin_class, adsimdet, fname):
     assert fname in str(lfname), f"{lfname=}  {fname=}"
 
 
-def test_full_file_name_local(adsimdet, fname):
+def test_full_file_name_local(fname):
+    adsimdet = custom_ad_creator("hdf1")
     plugin = adsimdet.hdf1
     plugin.file_name.put(fname)
     plugin.file_path.put(f"{plugin.write_path_template}/")
@@ -205,23 +318,8 @@ def test_full_file_name_local(adsimdet, fname):
     assert str(lfname).find(fname) > 0
 
 
-def test_no_file_path(adsimdet, fname):
-    adsimdet.hdf1.file_name.put(fname)
-    adsimdet.hdf1.file_path.put("/no-such-path-exists")
-    time.sleep(0.1)
-    assert adsimdet.hdf1.file_path_exists.get() == 0
-
-    adsimdet.hdf1.file_path.put("")
-    time.sleep(0.1)
-    assert adsimdet.hdf1.file_path_exists.get() == 0
-
-    with pytest.raises(OSError) as exc:
-        adsimdet.stage()
-    assert str(exc.value).endswith(" does not exist on IOC.")
-    adsimdet.unstage()
-
-
-def test_file_numbering(adsimdet, fname):
+def test_file_numbering(fname):
+    adsimdet = custom_ad_creator("hdf1")
     plugin = adsimdet.hdf1
     plugin.file_name.put(fname)
     plugin.file_path.put(f"{plugin.write_path_template}/")
@@ -256,7 +354,9 @@ def test_file_numbering(adsimdet, fname):
 
 
 def test_capture_error_with_single_mode(adsimdet):
+    # adsimdet = custom_ad_creator("hdf1")
     plugin = adsimdet.hdf1
+    plugin.file_name.put("test")
 
     # test that plugin is in the wrong configuration for Single mode
     assert "capture" in plugin.stage_sigs
@@ -309,6 +409,112 @@ def test_plugin(Klass, setup, attrs, context, expected):
         knowns = dir(obj)
         for attr in attrs:
             assert attr in knowns  # hasattr() needs connection
+
+    if expected is not None:
+        assert expected in str(exinfo)
+
+
+@pytest.mark.parametrize(
+    "setup",
+    [
+        dict(
+            plugins=[
+                "cam",
+                dict(
+                    hdf1={
+                        "class": AD_EpicsFileNameHDF5Plugin,
+                        "read_path_template": "/tmp/docker_ioc/iocad/tmp/",
+                        "write_path_template": "/tmp/",
+                    }
+                ),
+            ]
+        ),
+    ],
+)
+@pytest.mark.parametrize(
+    "attrs, context, expected",
+    [
+        [
+            dict(
+                auto_save="No",
+                create_directory=0,
+                file_name="",
+                file_number=1,
+                file_path="",
+                file_template="",
+            ),
+            pytest.raises(ValueError),
+            "'ad:HDF1:FileName' is an empty string.",
+        ],
+        [
+            dict(
+                auto_save="No",
+                create_directory=-5,
+                file_template="",
+                file_name="test",
+                file_number=1,
+                file_path="",
+            ),
+            pytest.raises(ValueError),
+            "'ad:HDF1:FilePath' is an empty string.",
+        ],
+        [
+            dict(
+                auto_save="No",
+                create_directory=-5,
+                file_template="",
+                file_name="test",
+                file_number=1,
+                file_path="/tmp/",
+            ),
+            pytest.raises(ValueError),
+            "'ad:HDF1:FileTemplate' is an empty string.",
+        ],
+        [
+            dict(
+                auto_save="No",
+                create_directory=-5,
+                file_template="%s%s_%6.6d.h5",
+                file_name="test",
+                file_number=1,
+                file_path="/tmp/",
+            ),
+            does_not_raise(),
+            None,
+        ],
+        [
+            dict(
+                auto_save="No",
+                create_directory=-5,
+                file_template="%s%s_%6.6d.h5",
+                file_name="test",
+                file_number=1,
+                file_path="/",
+            ),
+            pytest.raises(OSError),
+            "Path '/' does not exist on IOC",
+        ],
+    ],
+)
+def test_HDF5plugin_i1062(setup, attrs, context, expected):
+    """Issue #1062, Unconfigured HDF plugin."""
+    with context as exinfo:
+        det = ad_creator(IOC_AD, name="det", **setup)
+
+        # Confirm these attributes are not staged.
+        for attr in "file_name file_path file_template".split():
+            assert attr not in det.hdf1.stage_sigs
+
+        if not AD_plugin_primed(det.hdf1):
+            AD_prime_plugin2(det.hdf1)
+            assert AD_plugin_primed(det.hdf1)
+
+        for attr, value in attrs.items():
+            caput_AD(getattr(det.hdf1, attr), value)
+
+        RE = bluesky.RunEngine()
+        (uid,) = RE(bp.count([det]))
+        assert isinstance(uid, str)
 
     if expected is not None:
         assert expected in str(exinfo)